refactor(web): extract usePersonTracks hook from App.jsx (sc-1651)
Fifth Phase-A slice. Moves personTracks state + refreshPersonTracks,
createPersonDetectionJob, createPersonTrackJob, saveTrackCorrections into
hooks/usePersonTracks.js. personTracks is project-scoped (loaded by the project
effect, not the bulk refreshData), so the hook is self-contained like
useCharacters; detection/track jobs refresh via the injected refreshData. App
composes useTraining-style deps ({ token, activeProject, setError, requestedGpu,
setActiveView, refreshData }) and forwards values down unchanged, so VideoStudio
is untouched. personReadiness (derived from workers) stays in App.

Behavior-preserving: App.jsx 1893 -> 1820 lines (god module 2334 -> 1820, -514 over
five slices). 107 web vitest pass, build clean, browser smoke (Rust API on a temp
data dir): VideoStudio renders with all person-track props wired, the Replace-person
panel renders, no console errors. (Detection/track mutations need a GPU worker, so
wiring is verified by render.)
